Some more Wwise functions (not functional)
I spent 4+ hours on this, following the documentation just for their to be a parameter which I don't know the type of. Also the names in x32dbg were &"w_='" so I don't know how to do that. I tried escaping the string and putting it in. I honestly don't know why this doesn't work :-1: