Begin Linux Support
Got it to not crash by getting a place where I can hook, and removing some problematic hooks.

Only downside is that we need VertDecl for NumElements, VertShaderF for StartReg and VectCount, and Reset for ImGUI  / Font resets. RIP ThiccMesh (for now at least)

ER makes all my strings the low E's color, so that's cool.