Test Whammy DT multi-down tune
With @PoizenJam's work on Whammy / Bass Whammy, I wanna see if I can get the Whammy DT moved over and able to use it to take advantage of tunings that aren't supported in the "Shift Up" / "Shift Down" sections. This includes -8, -9, -10, -11, the positive counterparts, and potentially tunings above / below an octave.