Check For Updates -> Button
Instead of doing it automatically, we now just put a button saying that an update is available. It'll only show up IF an update is available, and will be hidden otherwise. I also moved a lot of the code into separate functions to call it when needed.

Tried with ASYNC but it ended up locking up the app, so I moved way from it.