Cleaned up looping logic more and made it play more friendly with the riff repeater.
- Validation of loopStart and loopEnd is done when they are initially set so it doesn't need to be checked again.
- Loop counter is ignored in the riff repeater (riff repeater functions normally even with a loop enabled)

There is still one very tricky bug I have not been able to resolve. When pressing "resume" from the riff repeater menu when you have a riff repeater section outside of the loop selected it does not properly seek back to the start of the loop and some notes can appear greyed out. Simply pausing and then resuming again (with ESC) resolves the bug but I would like to get this to be automatic.