Add a check for ImGui Init before running ImGui Shutdown
If the user opens Rocksmith with a RSMods DEBUG dll, and closes Rocksmith by closing the console terminal before the Hook_EndScene() is run, then an assert would be thrown when the user tries to close Rocksmith.

This issue came up when we updated ImGui a couple months back and only affects DEBUG versions of RSMods.