Allow Complex Uplay Passwords
Password requirements for Ubisoft Connect and Rocksmith are slightly different. Rocksmith restricts the use of 4 characters: " / \ and =. If a user has any of these characters in their Uplay password, they have to change their password to log into Rocksmith.

This mod fixes that by making it not restrict those characters.